# pkg/server/route/docs/swagger.yaml
definitions:
...
  v3.HttpReq_Cluster_create:
    properties:
      name:
        type: string
      polling_limit:
        type: integer
      polling_option:
        type: object
      summary:
        type: string
      uuid:
        description: uuid
        type: string
    type: object
  v3.HttpReq_Cluster_update:
    properties:
      name:
        type: string
      polling_limit:
        type: integer
      polling_option:
        type: object
      summary:
        type: string
    type: object
...
  v3.HttpRsp_Cluster:
    properties:
      created:
        type: string
      deleted:
        type: string
      id:
        description: pk
        type: integer
      name:
        type: string
      polling_limit:
        type: integer
      polling_option:
        type: object
      summary:
        type: string
      updated:
        type: string
      uuid:
        description: uuid
        type: string
    type: object
...
  v3.RegularPollingOption:
    type: object
...
  v3.SmartPollingOption:
    properties:
      busy:
        description: (초)
        type: integer
      idle:
        description: (초)
        type: integer
    type: object
info:
  contact:
    email: jaehoon@nexclipper.io
    url: https://nexclipper.io
  description: this is a sudory server.
  title: SUDORY
  version: 0.0.1
paths:
...
  /server/cluster:
    get:
      consumes:
      - application/json
      description: Find cluster
      parameters:
      - description: query  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: q
        type: string
      - description: order  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: o
        type: string
      - description: paging github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: p
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v3.HttpRsp_Cluster'
            type: array
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
    post:
      consumes:
      - application/json
      description: Create a cluster
      parameters:
      - description: HttpReq_Cluster_create
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/v3.HttpReq_Cluster_create'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.HttpRsp_Cluster'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
  /server/cluster/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete a cluster
      parameters:
      - description: Cluster Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
    get:
      consumes:
      - application/json
      description: Get a cluster
      parameters:
      - description: Cluster Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.HttpRsp_Cluster'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
    put:
      consumes:
      - application/json
      description: Update a cluster
      parameters:
      - description: Cluster Uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: HttpReq_Cluster_update
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/v3.HttpReq_Cluster_update'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.HttpRsp_Cluster'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
  /server/cluster/{uuid}/polling/regular:
    put:
      consumes:
      - application/json
      description: Update a cluster Polling Reguar
      parameters:
      - description: Cluster Uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: RegularPollingOption
        in: body
        name: polling_option
        required: true
        schema:
          $ref: '#/definitions/v3.RegularPollingOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.HttpRsp_Cluster'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
  /server/cluster/{uuid}/polling/smart:
    put:
      consumes:
      - application/json
      description: Update a cluster Polling Smart
      parameters:
      - description: Cluster Uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: SmartPollingOption
        in: body
        name: polling_option
        required: true
        schema:
          $ref: '#/definitions/v3.SmartPollingOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.HttpRsp_Cluster'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster
...
securityDefinitions:
  ClientAuth:
    description: token for client api
    in: header
    name: x-sudory-client-token
    type: apiKey
  ServiceAuth:
    description: Bearer token for service api
    in: header
    name: Authorization
    type: apiKey
  XAuthToken:
    description: limit for access sudory api
    in: header
    name: x_auth_token
    type: apiKey
swagger: "2.0"